Refactor AgentCore Lambda deployment for better modularity and reusability
Extract reusable components from monolithic files into focused modules:
- agent/config.py: TAC configuration loader with Secrets Manager integration
- agent/tac_app.py: AgentCore app adapter for HTTP/WebSocket protocols
- agent/websocket_adapter.py: WebSocket wrapper with welcome greeting support
- lambda/lambda_proxy.py: Reusable Lambda proxy class for Twilio webhooks

Benefits:
- Reduced main.py from 180 lines to 63 lines (65% reduction)
- Reduced index.py from 207 lines to 47 lines (77% reduction)
- Improved testability via dependency injection
- Better separation of concerns
- Easier to reuse across different Lambda functions
